+++ core/src/main/java/io/github/scroojalix/npcmanager/NPCMain.java
@@ -1,6 +1,5 @@
 package io.github.scroojalix.npcmanager;
 
-import java.sql.SQLException;
 import java.util.logging.Level;
 
 import org.bukkit.Bukkit;
@@ -13,11 +12,11 @@ import io.github.scroojalix.npcmanager.events.NPCEvents;
 import io.github.scroojalix.npcmanager.nms.interfaces.INPCManager;
 import io.github.scroojalix.npcmanager.nms.interfaces.IPacketReader;
 import io.github.scroojalix.npcmanager.utils.PluginUtils;
-import io.github.scroojalix.npcmanager.utils.PluginUtils.SaveMethod;
 import io.github.scroojalix.npcmanager.utils.PluginUtils.ServerVersion;
-import io.github.scroojalix.npcmanager.utils.chat.Messages;
 import io.github.scroojalix.npcmanager.utils.npc.equipment.EmptySlots;
-import io.github.scroojalix.npcmanager.utils.sql.MySQL;
+import io.github.scroojalix.npcmanager.utils.npc.equipment.EquipmentInventory;
+import io.github.scroojalix.npcmanager.utils.storage.Storage;
+import io.github.scroojalix.npcmanager.utils.storage.StorageFactory;
 
 /**
  * Main class for the NPCManager plugin.
@@ -32,8 +31,7 @@ public class NPCMain extends JavaPlugin {
 	
 	public INPCManager npc;
 	public IPacketReader reader;
-	public SaveMethod saveMethod;
-	public MySQL sql;
+	public Storage storage;
 	public boolean showDebugMessages;
 	
 	private boolean validVersion = true;
@@ -42,7 +40,6 @@ public class NPCMain extends JavaPlugin {
 	public void onEnable() {
 		NPCMain.instance = this;
 		this.showDebugMessages = this.getConfig().getBoolean("show-debug-messages");
-		this.setSaveMethod();
 		if (!validVersion()) {
 			getLogger().severe("This plugin is not compatible with that server version!");
 			getLogger().severe("Disabling the plugin.");
@@ -63,9 +60,20 @@ public class NPCMain extends JavaPlugin {
 	 * Initialises the plugin.
 	 */
 	private void initialise() {
+		String pack = "io.github.scroojalix.npcmanager.nms."+serverVersion;
+		try {
+			npc = (INPCManager) Class.forName(pack + ".NPCManager").getConstructors()[0].newInstance(this);
+			reader = (IPacketReader) Class.forName(pack + ".PacketReader").getConstructors()[0].newInstance(this);
+		} catch (Exception e) {
+			this.getLogger().log(Level.SEVERE, "Could not initialise the plugin", e);
+			validVersion = false;
+			this.setEnabled(false);
+			return;
+		}
 		this.saveDefaultConfig();
 		EmptySlots.generateItems();
-		this.initSaveMethod();
+		this.storage = new StorageFactory(this).getInstance();
+		this.storage.init();
 		this.getCommand("npc").setExecutor(new CommandManager(this));
 		this.getServer().getPluginManager().registerEvents(new NPCEvents(this), this);
 		this.getServer().getPluginManager().registerEvents(new EquipmentEvents(this), this);
@@ -79,12 +87,11 @@ public class NPCMain extends JavaPlugin {
 				npc.removeAllNPCs();
 			}
 			for (Player player : Bukkit.getOnlinePlayers()) {
-				player.closeInventory();
+				if (player.getOpenInventory().getTopInventory().getHolder() instanceof EquipmentInventory) 
+					player.closeInventory();
 				reader.uninject(player);
 			}
-			if (saveMethod == SaveMethod.MYSQL) {
-				sql.disconnect();
-			}
+			storage.shutdown();
 		}
 		instance = null;
 	}
@@ -94,25 +101,16 @@ public class NPCMain extends JavaPlugin {
 		try {
 			version = Bukkit.getServer().getClass().getPackage().getName().split("\\.")[3];
 		} catch (ArrayIndexOutOfBoundsException whatVersionAreYouUsingException) {
-			getLogger().severe("Unknown Server Version");
+			getLogger().log(Level.SEVERE, "Unknown Server Version", whatVersionAreYouUsingException);
 			return false;
 		}
 
 		try {
 			serverVersion = ServerVersion.valueOf(version);
 		} catch(Exception e) {
-			e.printStackTrace();
 			return false;
 		}
-
-		String pack = "io.github.scroojalix.npcmanager.nms."+version;
-		try {
-			npc = (INPCManager) Class.forName(pack + ".NPCManager").getConstructors()[0].newInstance(this);
-			reader = (IPacketReader) Class.forName(pack + ".PacketReader").getConstructors()[0].newInstance(this);
-		} catch (Exception e) {
-			return false;
-		}
-		return npc != null && reader != null;
+		return true;
 	}
 	
 	public void log(Level level, String msg) {
@@ -122,61 +120,17 @@ public class NPCMain extends JavaPlugin {
 	}
 	
 	public void reloadPlugin() {
+		for (Player player : Bukkit.getOnlinePlayers()) {
+			if (player.getOpenInventory().getTopInventory().getHolder() instanceof EquipmentInventory) 
+				player.closeInventory();
+		}
 		npc.removeAllNPCs();
 		this.saveDefaultConfig();
 		reloadConfig();
 		showDebugMessages = getConfig().getBoolean("show-debug-messages");
-		setSaveMethod();
-		initSaveMethod();
+		storage.shutdown();
+		this.storage = new StorageFactory(this).getInstance();
+		this.storage.init();
 		PluginUtils.checkForUpdate();
-	}
-	
-	private void setSaveMethod() {
-		String save = getConfig().getString("save-method");
-		switch(save) {
-		case "JSON":
-			saveMethod = SaveMethod.JSON;
-			log(Level.INFO, "Save method set to JSON");
-			break;
-		case "MYSQL":
-			saveMethod = SaveMethod.MYSQL;
-			log(Level.INFO, "Save method set to MYSQL");
-			break;
-		default:
-			log(Level.WARNING, "Unknown saving method '"+save+"'. Defaulting it to JSON");
-			saveMethod = SaveMethod.JSON;
-			break;
-		}
-	}
-
-	private void initSaveMethod() {
-		Bukkit.getScheduler().runTaskAsynchronously(this, new Runnable() {
-			@Override
-			public void run() {
-				switch(saveMethod) {
-				case JSON:
-					break;
-				case MYSQL:
-					NPCMain.instance.sql = new MySQL(NPCMain.instance);
-					try {
-						sql.connect();
-						if (sql.isConnected()) {
-							log(Level.INFO, "Successfully connected to database.");
-							sql.getGetter().createTable();
-						}
-					} catch (ClassNotFoundException | SQLException e) {
-						getLogger().log(Level.SEVERE, Messages.DATABASE_NOT_CONNECTED);
-					}
-					break;
-				}
-
-				Bukkit.getScheduler().runTask(NPCMain.instance, new Runnable() {
-					@Override
-					public void run() {
-						NPCMain.instance.npc.restoreNPCs();
-					}
-				});
-			}
-		});
 	}
 }