+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/modifications/InteractEventModification.java
@@ -48,14 +48,14 @@ public class InteractEventModification extends SubCommand {
                     command += " " + args[arg];
                 }
                 data.setInteractEvent(new CommandInteraction(command));
-                main.npc.saveNPC(data);
+                main.storage.saveNPC(data);
                 sender.sendMessage(PluginUtils.format(
                         "&6Set &F" + data.getName() + "'s &6Interact Event to the command &F/" + command));
                 return true;
             } else if (args[3].equalsIgnoreCase("custom") && args.length > 4) {
                 if (InteractionsManager.getInteractEvents().containsKey(args[4])) {
                     data.setInteractEvent(InteractionsManager.getInteractEvents().get(args[4]));
-                    main.npc.saveNPC(data);
+                    main.storage.saveNPC(data);
                     sender.sendMessage(PluginUtils
                             .format("&6Set &F" + data.getName() + "'s &6Interact Event to &F" + args[3]));
                 } else {
@@ -65,7 +65,7 @@ public class InteractEventModification extends SubCommand {
                 return true;
             } else if (args[3].equalsIgnoreCase("none")) {
                 data.setInteractEvent(null);
-                main.npc.saveNPC(data);
+                main.storage.saveNPC(data);
                 sender.sendMessage(PluginUtils.format("&6Removed the Interact Event for &F" + data.getName()));
                 return true;                
             }