+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/modifications/SkinModification.java
@@ -79,7 +79,7 @@ public class SkinModification extends SubCommand {
             } else if (args[3].equalsIgnoreCase("default")) {
                 data.getTraits().setSkinData(null);
                 data.getTraits().setSkinLayers(null);
-                main.npc.saveNPC(data);
+                main.storage.saveNPC(data);
                 main.npc.updateNPC(data);
                 sender.sendMessage(PluginUtils.format("&6Reset the skin of &F"+data.getName()));
                 return true;