+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/modifications/SubtitleModification.java
@@ -42,7 +42,7 @@ public class SubtitleModification extends SubCommand {
             }
             data.getTraits().setSubtitle(value);
             sender.sendMessage(PluginUtils.format("&6Set the subtitle of &F"+data.getName()+"&6 to &F"+value));
-            main.npc.saveNPC(data);
+            main.storage.saveNPC(data);
             main.npc.updateNPC(data);
             return true;
         }