+++ core/src/main/java/io/github/scroojalix/npcmanager/utils/storage/implementation/StorageImplementation.java
@@ -1,0 +1,17 @@
+package io.github.scroojalix.npcmanager.utils.storage.implementation;
+
+import io.github.scroojalix.npcmanager.utils.npc.NPCData;
+
+public interface StorageImplementation {
+    
+    public void init() throws Exception;
+
+    public void shutdown() throws Exception;
+
+    public void saveNPC(NPCData data);
+
+    public void removeNPC(String name);
+    
+    public void restoreNPCs();
+
+}