+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project uses the versioning format {phase}.{major}.{minor}, where phase is 
 - The type of Skin (URL/PLAYER) and UUID is now stored with the skin to make it easier to determine if a skin needs updating.
 - Can now reload a specific by running `/npc reload <npc>`.
 - Some of the confirmation messages. eg. `Removed an NPC named <npc>` -> `Removed <npc>`.
+- Better implementation for storage methods, making it easier to add new storage methods.
 ### Fixed
 - When the distance to an NPC is 0, the head rotation was weird.
 - Issue with modify command displaying all modifications if the input NPC doesn't exist.